Not Working PWM. Fights with either > 80% and < 30%.
Need to switch back to SPI. PWM driver does not work for both sides of
getting close to 75%, and 25% within the same code. and use DMA. SPI at
least will hopefully be able to provide less data consumption as well

+++ TestSoftware/SegmentDriver/SegmentDriver/Timer.h
@@ -2,4 +2,4 @@
 #include <stdint.h>
 
 void InitTimer();
-void SendTimerDMA(uint8_t* buffer, uint16_t count);
+void SendTimerDMA(uint32_t* buffer, uint16_t count);
